Working Bad Apple animation, no sound yet
